==> 2.preprocessing.ipynb <==
# Генерация фич и текстовый препроцессинг

## Общее описание
Ноутбук выполняет комплексную подготовку данных для задачи машинного обучения:
- Генерация новых признаков
- Текстовый препроцессинг
- Оптимизация типов данных
- Очистка от нерелевантных признаков

## Результирующие признаки

### Числовые признаки
| Признак | Описание | Тип данных |
|---------|----------|------------|
| `price_diff_log` | Логарифмированная разница цен (log1p) | float32 |
| `price_diff_is_outlier` | Флаг выбросов (метод IQR) | bool |
| `param1` | Совпадение значений параметра | int8 |

### Текстовые признаки
| Признак | Описание | Тип данных |
|---------|----------|------------|
| `base_text` | Конкатенация: base_title + base_description | object |
| `cand_text` | Конкатенация: cand_title + cand_description | object |

### Удаленные признаки
- Исходные текстовые поля (title/description)
- Ценовые показатели (base_price, cand_price)
- Категориальные признаки
- Параметры объявлений
- Географические признаки
- Мета-информация (images, json_params)

import pandas as pd
import numpy as np

import re

import warnings
warnings.filterwarnings('ignore')

# Загружаем отдельно test и train часть для обработки

train = pd.read_parquet('../data/train.parquet')

test = pd.read_parquet('../data/test.parquet')

test.info()

# Function

def some_diff(data: pd.DataFrame, 
              col1: str, 
              col2: str,
              name: str) -> pd.DataFrame:
    """Считает разность значений в двух столбцах
    :param data: датасет
    :param col1: название столбца 1
    :param col2: название столбца 2
    :param name: название нового столбца
    :return: датасет с новым столбцом
    """
    data[name] = abs(data[col1] - data[col2])
    
    return data

def price(data: pd.DataFrame) -> pd.DataFrame:
    """Преобразует признак цены"""
    data = some_diff(data, 'base_price', 'cand_price', 'price_diff')
    
    # прологарифмируем фичу, так как присутствуют сильные выбросы (log(1 + x))
    data['price_diff_log'] = np.log1p(data['price_diff'])
    
    # создадим флаг выброса как фичу
    Q1 = data['price_diff_log'].quantile(0.25)
    Q3 = data['price_diff_log'].quantile(0.75)
    IQR = Q3 - Q1

    # нижняя и верхняя границы для определения выбросов
    lower_bound = Q1 - 1.5 * IQR
    upper_bound = Q3 + 1.5 * IQR

    # флаг выброса
    data['price_diff_is_outlier'] = (data['price_diff_log'] < lower_bound) | (
        data['price_diff_log'] > upper_bound)
    
    # преобразуем в оптимальных тип данных
    data['price_diff_log'] = data['price_diff_log'].astype('float32')
    
    return data

def combine_columns(data: pd.DataFrame) -> pd.DataFrame:
    """Соединяет текстовые фичи в новую и удаляет старые столбцы"""
    data['base_text'] = data['base_title'].astype(
        str) + '. ' + data['base_description'].astype(str)
    data['cand_text'] = data['cand_title'].astype(
        str) + '. ' + data['cand_description'].astype(str)
    
    return data

def cleaned_text(text: str) -> str:
    """Простая очистка текста"""
    # преобразуем text в строку, если это не строка
    text = str(text) if text is not None else ''
    text = text.lower()
    text = re.sub(r'[^а-яёa-z0-9\s.,*!?:-]', '', text) # удаление лишних символов (кроме пунктуации)
    text = re.sub(r'\s+', ' ', text).strip() # удаление лишних пробелов
    text = re.sub(r'^[^\w]+', '', text) # удаление пунктуации в начале строки
    text = text.strip(' .') # убирает лишние пробелы и точки в начале и в конце строки
    
    return text

# Preprocessing

def preprocessing(data: pd.DataFrame) -> pd.DataFrame:
    """
    Препроцессинг датасета
    """
    # получаем признаки из цены
    data = price(data)

    # получаем фичу param1
    data['param1'] = (data['base_param1'].astype(str) ==
                      data['cand_param1'].astype(str)).astype(int)

    # получаем текстовые признаки
    data = combine_columns(data)

    # удаляем столбцы датасета
    data.drop([
        'base_title', 'base_description', 'cand_title', 'cand_description',
        'base_price', 'cand_price', 'price_diff', 'base_category_name',
        'cand_category_name', 'base_subcategory_name', 'cand_subcategory_name',
        'base_param1', 'cand_param1', 'base_param2', 'cand_param2',
        'base_count_images', 'cand_count_images', 'base_json_params',
        'cand_json_params', 'is_same_location', 'is_same_region',
        'base_count_images', 'cand_count_images'
    ],
              axis=1,
              inplace=True)

    # преобразовываем данные в оптимальный тип
    to_int32 = ['group_id']
    to_int8 = ['is_double', 'param1', 'count_images_diff']

    for col in data.columns:
        if data[col].dtype == 'int64' and col in to_int32:
            data[col] = data[col].astype('int32')
        elif data[col].dtype == 'int64' and col in to_int8:
            data[col] = data[col].astype('int8')

    # чистка текста
    for i in data.columns:
        if data[i].dtype == 'object':
            data[i] = data[i].apply(cleaned_text)
            data[i] = data[i].astype('object')
        elif data[i].dtype == 'category':
            data[i] = data[i].apply(cleaned_text)
            data[i] = data[i].astype('category')

    return data

## Test

df_test = preprocessing(test)
df_test.head()

df_test.info()

# сохраняем предобработанный датасет
#df_test.to_parquet('../data/df_test.parquet', index=False)

## Train

df_train = preprocessing(train)

df_train.drop(['action_date', 'group_id'], axis=1, inplace=True)
df_train.rename(columns={'is_double': 'target'}, inplace=True) 
df_train['target'] = df_train.pop('target')  # переносит target в последнюю колонку

df_train.head()

df_train.info()

# сохраняем предобработанный датасет
#df_train.to_parquet('../data/df_train.parquet', index=False)

**Выводы:**
После просмотра и анализа тестовых данных, была немного изменена предобработка тренировочных данных. А именно из-за пропусков в столбцах количества изображений в тестовых данных, на трейне принято решение ее не использовать (по графикам в EDA также видно, что она несет много шума), также исходя из этого в трейне не были удалены строки с пропусками. На тесте отсутствует столбец action_date, следовательно фичи на его основе удалены из трейна. 

 
 Следующие шаги:
 1. На основе полученных эмбеддингов картинок и текстов, соответственно, посчитать их косинусную близость, разность и произведение.
 2. Из получившихся результатов разности и произведения соответствующих эмбеддингов, представить их в более низкоразмерном пространстве (например размерами 64 или 32 с помощью PCA или TNSE).
 3. Составить на основе всего вышеизложенного матрицу объект-признаки и обучить катбуст на задачу классификации.
