==> 3.text_embedding.ipynb <==
# Обработка эмбеддингов товаров с помощью rubert-tiny-turbo

Этот ноутбук предназначен для преобразования текстовых описаний товаров в векторные представления (эмбеддинги) с использованием предобученной модели `rubert-tiny-turbo`.
В ноутбуке представлен вариант для обработки заголовков (проверка гипотезы о том, что в них достаточно информации).

## Основные этапы обработки

1. **Загрузка данных**:
   - Отдельная обработка train и test выборок
   - Предварительный анализ текстовых описаний

2. **Генерация эмбеддингов**:
   - Использование модели `rubert-tiny-turbo` для получения векторных представлений
   - Оптимизированная обработка больших объемов текста

3. **Сохранение результатов**:
   - Формирование датафреймов с эмбеддингами
   - Сохранение в удобном формате для последующего использования


На выходе получаем:
- `df_train.parquet` - эмбеддинги для обучающей выборки
- `df_test.parquet` - эмбеддинги для тестовой выборки

from sentence_transformers import SentenceTransformer
import pandas as pd

train = pd.read_parquet('df_train.parquet')
test = pd.read_parquet('df_test.parquet')

train

model = SentenceTransformer('sergeyzh/rubert-tiny-turbo', device='cuda')

# Получение эмбеддингов для train и test части.

# for train part

import tqdm

# для df1
base_emb = []
cand_emb = []

for i in tqdm.tqdm(range(len(train))):

  sentences = [train['base_title'][i], train['cand_title'][i]]

  embeddings = model.encode(sentences, batch_size=128, convert_to_tensor=True)


  #embeddings = return_embeddings(df1['base_text'][i], df1['cand_text'][i])
  base_emb.append(embeddings[0].cpu().numpy())
  cand_emb.append(embeddings[1].cpu().numpy())

  

train['base_title_emb'] = base_emb
train['cand_title_emb'] = cand_emb

train

train.to_parquet('df_train.parquet', index=False)

## for test

import tqdm

# для df1
base_emb = []
cand_emb = []

for i in tqdm.tqdm(range(len(test))):

  sentences = [test['base_title'][i], test['cand_title'][i]]

  embeddings = model.encode(sentences, batch_size=128, convert_to_tensor=True)


  #embeddings = return_embeddings(df1['base_text'][i], df1['cand_text'][i])
  base_emb.append(embeddings[0].cpu().numpy())
  cand_emb.append(embeddings[1].cpu().numpy())


test['base_title_emb'] = base_emb
test['cand_title_emb'] = cand_emb

test.to_parquet('df_test.parquet', index=False)

test.info()
